+++ packages/basic-crawler/src/internals/basic-crawler.ts
@@ -635,6 +635,9 @@ export class BasicCrawler<Context extends CrawlingContext = BasicCrawlingContext
         // override the default internal timeout of request queue to respect `requestHandlerTimeoutMillis`
         if (this.requestQueue) {
             this.requestQueue.internalTimeoutMillis = this.internalTimeoutMillis;
+            // for request queue v2, we want to lock requests by the timeout provided by the user (plus 5 seconds padding), but
+            // with a minimum of a minute
+            this.requestQueue.requestLockSecs = Math.max(this.requestHandlerTimeoutMillis / 1000 + 5, 60);
         }
 
         this.maxRequestRetries = maxRequestRetries;
